+++ src/Controller/MessagesController.php
@@ -9,8 +9,13 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
+/**
+ * @security("is_granted('ROLE_EFFECTIF') or is_granted('ROLE_FAMILLE') or is_granted('ROLE_ADMIN')")
+ */
 class MessagesController extends AbstractController
 {
     /**
@@ -30,11 +35,14 @@ class MessagesController extends AbstractController
     /**
      * @Route("/{id}/messagesdelete", name="messagess_delete")
      */
-    public function delete(Messages $messages)
+    public function delete(Messages $messages, AuthorizationCheckerInterface $authChecker)
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($messages);
-        $entityManager->flush();
+        if ($messages->getUsers()->getId() == $this->user->getId() || true === $authChecker->isGranted('ROLE_ADMIN')) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($messages);
+            $entityManager->flush();
+        }else{
+        }
         return $this->redirectToRoute('forum');
     }
 
@@ -43,10 +51,14 @@ class MessagesController extends AbstractController
      * @Route("/{idS}/messagessadd", name="messages_add")
      * @Route("/{id}/messagesedit", name="messages_edit")
      */
-    public function add_edit(UserInterface $user, Messages $messages = null, Request $request, Sujets $sujet)
+    public function add_edit(UserInterface $user, AuthorizationCheckerInterface $authChecker, Messages $messages = null, Request $request, Sujets $sujet)
     {
+
+        if (($messages && $messages->getUsers()->getId() != $user->getId()) && false === $authChecker->isGranted('ROLE_ADMIN')) {     
+            return $this->redirectToRoute('forum');
+        }
         if (!$messages) {
-            $messages = new Messages($user , $sujet);
+            $messages = new Messages($user, $sujet);
         }
 
         $form = $this->createForm(MessagesType::class, $messages);
