+++ src/Controller/SujetsController.php
@@ -8,8 +8,13 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
+/**
+ * @security("is_granted('ROLE_EFFECTIF') or is_granted('ROLE_FAMILLE') or is_granted('ROLE_ADMIN') or is_granted('ROLE_ACCUEIL')")
+ */
 class SujetsController extends AbstractController
 {
     /**
@@ -29,11 +34,13 @@ class SujetsController extends AbstractController
     /**
      * @Route("/{id}/sujetsdelete", name="sujets_delete")
      */
-    public function delete(Sujets $sujets)
+    public function delete(Sujets $sujets, AuthorizationCheckerInterface $authChecker)
     {
+        if ($sujets->getUsers()->getId() == $this->user->getId() || true === $authChecker->isGranted('ROLE_ADMIN')) {
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($sujets);
-        $entityManager->flush();
+        $entityManager->flush();    
+        }
         return $this->redirectToRoute('forum');
     }
 
@@ -42,9 +49,11 @@ class SujetsController extends AbstractController
      * @Route("/sujetsadd", name="sujets_add")
      * @Route("/{id}/sujetsedit", name="sujets_edit")
      */
-    public function add_edit(UserInterface $user, Sujets $sujets = null, Request $request)
+    public function add_edit(UserInterface $user, Sujets $sujets = null, Request $request, AuthorizationCheckerInterface $authChecker)
     {
-
+        if (($sujets && $sujets->getUsers()->getId() != $user->getId()) && false === $authChecker->isGranted('ROLE_ADMIN')) {     
+            return $this->redirectToRoute('forum');
+        }
         if (!$sujets) {
             $sujets = new Sujets($user);
         }
